refactor: moved simulation screen to shared
